move all data out of resource
